api_yamdb/requirements.txt
attrs==22.1.0
blessed==1.19.1
bpython==0.23
certifi==2022.6.15
charset-normalizer==2.0.12
curtsies==0.4.0
cwcwidth==0.1.7
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
greenlet==1.1.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
gunicorn==20.0.4
packaging==21.3
pluggy==0.13.1
psycopg2-binary
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.2.1
pyxdg==0.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
wcwidth==0.2.5
